Splitup test
